refactor reserve and release endpoints in HTTP POST and DELETE methods
