+++ lib/db/api.js
@@ -121,7 +121,7 @@ dbapi.lookupUserByVncAuthResponse = function(response, serial) {
     })
 }
 
-dbapi.loadGroup = function(email) {
+dbapi.loadUserDevices = function(email) {
   return db.run(r.table('devices').getAll(email, {
     index: 'owner'
   }))
