+++ lib/units/api/controllers/device.js
@@ -1,18 +1,15 @@
 var Promise = require('bluebird')
+var _ = require('lodash')
 
 var dbapi = require('../../../db/api')
 var logger = require('../../../util/logger')
 var datautil = require('../../../util/datautil')
 
 var log = logger.createLogger('api:controllers:device')
-var wire = require('../../../wire')
-var wireutil = require('../../../wire/util')
 
 module.exports = {
   getDevices: getDevices
 , getDeviceBySerial: getDeviceBySerial
-, reserveDeviceBySerial: reserveDeviceBySerial
-, releaseDeviceBySerial: releaseDeviceBySerial
 }
 
 var log = logger.createLogger('api:contoller:device')
@@ -42,11 +39,17 @@ function getDevices(req, res) {
 
 function getDeviceBySerial(req, res) {
   var serial = req.swagger.params.serial.value
+  var fields = req.swagger.params.fields.value
 
   dbapi.loadDevice(serial)
     .then(function(device) {
       if (device) {
         datautil.normalize(device, req.user)
+
+        if(fields) {
+          device = _.pick(device, fields.split(','))
+        }
+
         res.json({
           success: true
         , device: device
@@ -63,101 +66,5 @@ function getDeviceBySerial(req, res) {
       res.json(500, {
         success: false
       })
-    })
-}
-
-function reserveDeviceBySerial(req, res) {
-  var serial = req.swagger.params.serial.value
-
-  dbapi.loadDevice(serial)
-    .then(function(device) {
-      if (device) {
-        datautil.normalize(device, req.user)
-        if(!device.using) {
-
-          var requirements = {
-            'serial': {
-              'value': serial,
-              'match': 'exact'
-            }
-          }
-
-          req.options.push.send([
-            device.channel
-          , wireutil.envelope(
-              new wire.GroupMessage(
-                new wire.OwnerMessage(
-                  req.user.email
-                , req.user.name
-                , req.user.group
-                )
-              , null
-              , wireutil.toDeviceRequirements(requirements)
-              )
-            )
-          ])
-
-          res.json({
-            success: true
-          , device: device
-          })
-
-        } else {
-          res.json(500, {
-            success: false
-          , description: 'Device is being used'
-          })
-        }
-      } else {
-        res.json(500, {
-          success: false
-        , description: 'Bad device serial'
-        })
-      }
-    })
-}
-
-function releaseDeviceBySerial(req, res) {
-  var serial = req.swagger.params.serial.value
-
-  dbapi.loadDevice(serial)
-    .then(function(device) {
-      if (device) {
-        datautil.normalize(device, req.user)
-        if(device.using && device.owner.email == req.user.email) {
-
-          var requirements = {
-            'serial': {
-              'value': serial,
-              'match': 'exact'
-            }
-          }
-
-          req.options.push.send([
-            device.channel
-          , wireutil.envelope(
-              new wire.UngroupMessage(
-                wireutil.toDeviceRequirements(requirements)
-              )
-            )
-          ])
-
-          res.json({
-            success: true
-          , device: device
-          })
-
-        } else {
-          res.json(500, {
-            success: false
-          , description: 'You cannot kick this device'
-          })
-        }
-      } else {
-        res.json(500, {
-          success: false
-        , description: 'Bad device serial'
-        })
-      }
     })
 }
