+++ lib/units/api/controllers/user.js
@@ -3,12 +3,16 @@ var Promise = require('bluebird')
 var dbapi = require('../../../db/api')
 var logger = require('../../../util/logger')
 var datautil = require('../../../util/datautil')
+var wire = require('../../../wire')
+var wireutil = require('../../../wire/util')
 
 var log = logger.createLogger('api:controllers:user')
 
 module.exports = {
   getCurrentUser: getCurrentUser
-, getCurrentUserGroup: getCurrentUserGroup
+, getCurrentUserDevices: getCurrentUserDevices
+, addDeviceToUser: addDeviceToUser
+, deleteDeviceFromUser: deleteDeviceFromUser
 }
 
 function getCurrentUser(req, res) {
@@ -18,8 +22,8 @@ function getCurrentUser(req, res) {
   })
 }
 
-function getCurrentUserGroup(req, res) {
-  dbapi.loadGroup(req.user.email)
+function getCurrentUserDevices(req, res) {
+  dbapi.loadUserDevices(req.user.email)
     .then(function(cursor) {
       return Promise.promisify(cursor.toArray, cursor)()
         .then(function(list) {
@@ -37,5 +41,102 @@ function getCurrentUserGroup(req, res) {
       res.json(500, {
         success: false
       })
+    })
+}
+
+function addDeviceToUser(req, res) {
+  var serial = req.body.serial
+  var timeout = req.body.timeout || null
+
+  dbapi.loadDevice(serial)
+    .then(function(device) {
+      if (device) {
+        datautil.normalize(device, req.user)
+        if(device.ready && !device.using) {
+
+          var requirements = {
+            'serial': {
+              'value': serial,
+              'match': 'exact'
+            }
+          }
+
+          req.options.push.send([
+            device.channel
+          , wireutil.envelope(
+              new wire.GroupMessage(
+                new wire.OwnerMessage(
+                  req.user.email
+                , req.user.name
+                , req.user.group
+                )
+              , timeout
+              , wireutil.toDeviceRequirements(requirements)
+              )
+            )
+          ])
+
+          res.json(202, {
+            success: true
+          , description: 'Device Add request is accepted'
+          })
+
+        } else {
+          res.json(500, {
+            success: false
+          , description: 'Device is being used or not available'
+          })
+        }
+      } else {
+        res.json(500, {
+          success: false
+        , description: 'Bad device serial'
+        })
+      }
+    })
+}
+
+function deleteDeviceFromUser(req, res) {
+  var serial = req.body.serial
+
+  dbapi.loadDevice(serial)
+    .then(function(device) {
+      if (device) {
+        datautil.normalize(device, req.user)
+        if(device.using && device.owner.email == req.user.email) {
+
+          var requirements = {
+            'serial': {
+              'value': serial,
+              'match': 'exact'
+            }
+          }
+
+          req.options.push.send([
+            device.channel
+          , wireutil.envelope(
+              new wire.UngroupMessage(
+                wireutil.toDeviceRequirements(requirements)
+              )
+            )
+          ])
+
+          res.json(202, {
+            success: true
+          , description: 'Device Release request is accepted'
+          })
+
+        } else {
+          res.json(500, {
+            success: false
+          , description: 'You cannot kick this device'
+          })
+        }
+      } else {
+        res.json(500, {
+          success: false
+        , description: 'Bad device serial'
+        })
+      }
     })
 }
